% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msentropy.R
\name{calculate_entropy_similarity}
\alias{calculate_entropy_similarity}
\title{Entropy similarity between two spectra}
\usage{
calculate_entropy_similarity(
  spMat1,
  spMat2,
  ms2_tolerance_in_da = -1,
  ms2_tolerance_in_ppm = 5
)
}
\arguments{
\item{spMat1}{A matrix of spectral peaks, with two columns: mz and intensity}

\item{spMat2}{A matrix of spectral peaks, with two columns: mz and intensity}

\item{ms2_tolerance_in_da}{The MS2 tolerance in Da, set to -1 to disable}

\item{ms2_tolerance_in_ppm}{The MS2 tolerance in ppm, set to -1 to disable}
}
\value{
The entropy similarity
}
\description{
Calculate the entropy similarity between two spectra
Two spectra sholud be cleaned before calculating.
}
\examples{
mz_a <- c(169.071, 186.066, 186.0769)
intensity_a <- c(7.917962, 1.021589, 100.0)
mz_b <- c(120.212, 169.071, 186.066)
intensity_b <- c(37.16, 66.83, 999.0)
spMat1 <- matrix(c(mz_a, intensity_a), ncol = 2, byrow = FALSE)
spMat2 <- matrix(c(mz_b, intensity_b), ncol = 2, byrow = FALSE)
calculate_entropy_similarity(spMat1, spMat2, ppm = 5)
}
