% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msentropy.R
\name{calculate_unweighted_entropy_similarity}
\alias{calculate_unweighted_entropy_similarity}
\title{Unweighted entropy similarity between two spectra}
\usage{
calculate_unweighted_entropy_similarity(spMat1, spMat2, ppm = 5)
}
\arguments{
\item{spMat1}{A matrix of spectral peaks, with two columns: mz and intensity}

\item{spMat2}{A matrix of spectral peaks, with two columns: mz and intensity}

\item{ppm}{The MS2 tolerance in ppm, set to -1 to disable}
}
\value{
The unweighted entropy similarity
}
\description{
Calculate the unweighted entropy similarity between two spectra.
Two spectra sholud be cleaned before calculating.
}
\examples{
mz_a <- c(169.071, 186.066, 186.0769)
intensity_a <- c(7.917962, 1.021589, 100.0)
mz_b <- c(120.212, 169.071, 186.066)
intensity_b <- c(37.16, 66.83, 999.0)
spMat1 <- matrix(c(mz_a, intensity_a), ncol = 2, byrow = FALSE)
spMat2 <- matrix(c(mz_b, intensity_b), ncol = 2, byrow = FALSE)
calculate_unweighted_entropy_similarity(spMat1, spMat2, ppm = 5)
}
