% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_matrix.R
\name{compare_spMat_entropy}
\alias{compare_spMat_entropy}
\title{Compare spMat entropy}
\usage{
compare_spMat_entropy(
  x,
  y,
  ms2_tolerance_in_da = -1,
  ms2_tolerance_in_ppm = 5,
  unweighted = FALSE
)
}
\arguments{
\item{x}{Query spMat}

\item{y}{Library spMat}

\item{ms2_tolerance_in_da}{The MS2 tolerance in Da, set to -1 to disable}

\item{ms2_tolerance_in_ppm}{The MS2 tolerance in ppm, set to -1 to disable}
}
\value{
The entropy similarity
}
\description{
Compare two spMat using entropy
}
\examples{
mz <- c(21.3300, 40.1320, 86.3400, 138.3290, 276.5710, 276.5830)
intensity <- c(100, 1300, 4030, 10000, 31600, 1000)
standardRow1 <- tibble::tibble(mz = list(mz), intensity = list(intensity))
spMat1 <- get_spMat(standardRow1)
mz <- c(21.3000, 40.1120, 86.3200, 138.3210, 276.5310)
intensity <- c(100, 1300, 4030, 10000, 31600)
standardRow2 <- tibble::tibble(mz = list(mz), intensity = list(intensity))
spMat2 <- get_spMat(standardRow2)
# Normalize intensity (sum to 1)
spMat1 <- clean_spMat(spMat1, normalize_intensity = TRUE)
spMat2 <- clean_spMat(spMat2, normalize_intensity = TRUE)
compare_spMat_entropy(x = spMat1, y = spMat2, ms2_tolerance_in_da = 0.02)
}
