% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_matrix.R
\name{get_spMat}
\alias{get_spMat}
\title{Get a spectrum matrix}
\usage{
get_spMat(standardRow)
}
\arguments{
\item{standardRow}{A tibble which need to include two columns mz and intensity.}
}
\value{
\code{matrix()} with two columns mz and intensity
}
\description{
Get a spectrum matrix from a tibble.
}
\examples{
mz <- c(40.1320, 86.3400, 138.3290, 276.5710)
intensity <- c(1300, 4030, 10000, 31600)
standardRow <- tibble::tibble(mz = list(mz), intensity = list(intensity))
get_spMat(standardRow)
}
