% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_spectra.R
\name{round_spMat}
\alias{round_spMat}
\alias{merge_spectra}
\title{Merge spectra}
\usage{
round_spMat(spMat, digits = 0)

merge_spectra(spMatList, digits = 0, method = sum)
}
\arguments{
\item{spMat}{spectra matrix with mz and intensity}

\item{digits}{mz digits}

\item{spMatList}{spMat list}

\item{method}{sum, median, mean}
}
\value{
spMat
}
\description{
When we are comparing multiple spectra that have inconsistent mz precision, we often need to
first change the mz precision to consistent. In addition, there are often multiple spectra for
one metabolite in the database, and combining these spectra is useful for metabolite identification.
}
\examples{
hmdbMsTb <- pubmsR::load_hmdbMsTb()
spMatList <- lapply(1:nrow(hmdbMsTb[hmdbMsTb$accession == "HMDB0000062", ]), function(i) {
  MetaboSpectra::get_spMat(hmdbMsTb[hmdbMsTb$accession == "HMDB0000062", ][i, ])
})
round_spMat(spMat = spMatList[[2]], digits = 0)
round_spMat(spMat = spMatList[[2]], digits = 1)
round_spMat(spMat = spMatList[[2]], digits = 2)
merge_spectra(spMatList = spMatList, method = sum)
}
