% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plotSpectra}
\alias{plotSpectra}
\title{plotSpectra}
\usage{
plotSpectra(spMat, num = 10, min_mz = NA, max_mz = NA)
}
\arguments{
\item{spMat}{spectra matrix}

\item{num}{max peak number}

\item{min_mz}{The minimum value of the horizontal coordinate}

\item{max_mz}{The maximum value of the horizontal coordinate}
}
\value{
A ggplot object
}
\description{
Plot a ggplot-style mass spectrum.
}
\examples{
mz <- c(40.1320, 86.3400, 138.3290, 276.5710)
intensity <- c(1300, 4030, 10000, 31600)
standardRow <- tibble::tibble(mz = list(mz), intensity = list(intensity))
spMat <- get_spMat(standardRow)
plotSpectra(spMat)
}
