% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_matrix.R
\name{searchLib_entropy}
\alias{searchLib_entropy}
\title{searchLib_entropy}
\usage{
searchLib_entropy(
  standardInput,
  lib,
  st = 0.8,
  tol_da1 = 0.01,
  tol_ppm1 = -1,
  tol_da2 = 0.02,
  tol_ppm2 = -1,
  predicted = "All",
  thread = 1,
  min_mz = -1,
  max_mz = -1,
  noise_threshold = 0.01,
  max_peak_num = -1
)
}
\arguments{
\item{standardInput}{A tibble, including six columns of information,
feature: name of the feature,
precursorMz: parent ion mz,
rt: retention time,
adduct: adduct form,
mz: mz of secondary mass spectrum,
intensity: intensity of secondary mass spectrum;
Both mz and intensity are stored in tibble as a list.}

\item{lib}{From any of the publicMs2List databases.}

\item{st}{Threshold for scoring.}

\item{tol_da1}{Under the Da unit, the tolerance of the same parent ion mz is searched.}

\item{tol_ppm1}{Under the ppm unit, the tolerance of the same parent ion mz is searched.}

\item{tol_da2}{Under the Da unit, two peaks are considered the tolerance of one peak.}

\item{tol_ppm2}{Under the ppm unit, two peaks are considered the tolerance of one peak.}

\item{predicted}{Whether to search only predictive mass spectra;All, Yes, No.}

\item{thread}{The number of threads running the function.}

\item{min_mz}{Minimum value of mz.}

\item{max_mz}{Maximum value of mz.}

\item{noise_threshold}{Peak intensity below noise_threshold * max(intensity) will be considered noise.}

\item{max_peak_num}{The number of peaks in the mass spectrum.}
}
\value{
searchRes_entropy, a list.
}
\description{
The mass spectrometry database was screened using the spectral entropy algorithm.
}
\details{
min_mz, max_mz, noise_threshold, max_peak_num, scale_int are parameter of function clean_spectra.
}
\examples{
data("standardInput", package = "MetaboSpectra")
load("publicMs2List.RData") # please see https://github.com/dongdongdong7/MetaboLib.ms2 and download publicMs2List.RData
searchRes_entropy <- searchLib_entropy(standardInput = standardInput, lib = publicMs2List$hmdb, thread = 8, st = 0.5)
}
