% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{searchRes2idenTibble}
\alias{searchRes2idenTibble}
\title{searchRes2idenTibble}
\usage{
searchRes2idenTibble(standardInput, searchRes, top = 5)
}
\arguments{
\item{standardInput}{A tibble, including six columns of information,
id: name of the feature,
precursorMz: parent ion mz,
rt: retention time,
adduct: adduct form,
mz: mz of secondary mass spectrum,
intensity: intensity of secondary mass spectrum;
Both mz and intensity are stored in tibble as a list.}

\item{searchRes}{Return of the searchLib function}

\item{top}{Retain the results of the top scores}
}
\value{
A tibble.
}
\description{
Change searchRes to idenTibble
}
\examples{
data("standardInput", package = "MetaboSpectra")
load("publicMs2List.RData") # please see https://github.com/dongdongdong7/MetaboLib.ms2 and download publicMs2List.RData
searchRes_entropy <- searchLib_entropy(standardInput = standardInput, lib = publicMs2List$hmdb, thread = 8, st = 0.5)
idenTibble_entropy <- searchRes2idenTibble(standardInput, searchRes_entropy, top = 5)
}
