% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim_peaksMatrix.R
\name{slim_peaksMatrix}
\alias{slim_peaksMatrix}
\alias{batch_slim_peaksMatrix}
\title{Slim peaks matrix (spectrum)}
\usage{
slim_peaksMatrix(
  peaksMatrix,
  ppm = 5,
  mz_method = c("median", "mean", "max_intensity"),
  intensity_method = c("sum", "mean", "max")
)

batch_slim_peaksMatrix(
  peaksMatrixList,
  ppm = 5,
  mz_method = c("median", "mean", "max_intensity"),
  intensity_method = c("sum", "mean", "max")
)
}
\arguments{
\item{peaksMatrix}{Peaks matrix (spectrum) from ms1 or ms2; also named spMat. It must be centroided}

\item{ppm}{The minimum mz difference in ppm to merge peaks}

\item{mz_method}{Method for calculating the merged peak mz}

\item{intensity_method}{Method for calculating the merged peak intensity}
}
\value{
peaksMatrix

peaksMatrixList
}
\description{
Merge adjacent peaks within the ppm range
}
\examples{
set.seed(123)
spMat <- cbind(
  mz = sort(rnorm(1000, mean = 100, sd = 0.1)),
  intensity = rgamma(1000, shape = 2, scale = 1)
)
system.time({
lapply(1:1000, function(i) {
  slim_peaksMatrix(peaksMatrix = spMat, ppm = 30)
 })
})
spMatList <- lapply(1:1000, function(i) {spMat})
system.time({
batch_slim_peaksMatrix(peaksMatrixList = spMatList)
})
}
