% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sps2dps.R
\name{sps2dps}
\alias{sps2dps}
\title{Convert Spectra to data points data.table}
\usage{
sps2dps(
  sps,
  ppm = 10,
  noise = 100,
  mz_method = "max_intensity",
  intensity_method = "sum"
)
}
\arguments{
\item{sps}{Spectra object for ms1 or ms2 from one sample.
It is reconmmended to first perform empty spectra filtering.}

\item{ppm}{The minimum mz difference in ppm to merge peaks}

\item{noise}{noise value of spectra}

\item{mz_method}{Method for calculating the merged peak mz}

\item{intensity_method}{Method for calculating the merged peak intensity}
}
\value{
data.table
}
\description{
Convert Spectra to data points data.table
}
