% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msentropy.R
\name{clean_spectrum}
\alias{clean_spectrum}
\title{Clean a spectrum}
\usage{
clean_spectrum(
  spMat,
  min_mz = -1,
  max_mz = -1,
  noise_threshold = 0.01,
  ppm = 5,
  max_peak_num = -1,
  normalize_intensity = FALSE
)
}
\arguments{
\item{spMat}{A matrix of spectrum, with two columns: m/z and intensity}

\item{min_mz}{The minimum mz value to keep, set to -1 to disable}

\item{max_mz}{The maximum mz value to keep, set to -1 to disable}

\item{noise_threshold}{The noise threshold, set to -1 to disable, all peaks have intensity < noise_threshold * max_intensity will be removed}

\item{ppm}{The minimum mz difference in ppm to merge peaks, any two peaks with mz difference < ppm will be merged}

\item{max_peak_num}{The maximum number of peaks to keep, set to -1 to disable}

\item{normalize_intensity}{Whether to normalize the intensity to sum to 1}
}
\value{
A matrix of spectral peaks, with two columns: mz and intensity
}
\description{
Clean a spectrum
}
\examples{
mz <- c(100.212, 169.071, 169.078, 300.321)
intensity <- c(0.3716, 7.917962, 100., 66.83)
spMat <- matrix(c(mz, intensity), ncol = 2, byrow = FALSE)
clean_spectrum(spMat, min_mz = 0, max_mz = 1000, noise_threshold = 0.01,
               ppm = -1,
               max_peak_num = 100, normalize_intensity = TRUE)
}
