% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_matrix.R
\name{clean_spMat}
\alias{clean_spMat}
\title{Clean spectrum matrix}
\usage{
clean_spMat(
  spMat,
  min_ms2_difference_in_da = -1,
  min_ms2_difference_in_ppm = 5,
  min_mz = -1,
  max_mz = -1,
  noise_threshold = 0.01,
  max_peak_num = -1,
  scale_int = 100,
  normalize_intensity = FALSE,
  precursorMz = NA
)
}
\arguments{
\item{spMat}{The matrix from the mass spectrum needs to contain two columns, mz and intensity.}

\item{min_ms2_difference_in_da}{The minimum mz difference in Da to merge peaks, set to -1 to disable, any two peaks with mz difference < min_ms2_difference_in_da will be merged}

\item{min_ms2_difference_in_ppm}{The minimum mz difference in ppm to merge peaks, set to -1 to disable, any two peaks with mz difference < min_ms2_difference_in_ppm will be merged}

\item{min_mz}{The minimum mz value to keep, set to -1 to disable}

\item{max_mz}{The maximum mz value to keep, set to -1 to disable}

\item{noise_threshold}{The noise threshold, set to -1 to disable, all peaks have intensity < noise_threshold * max_intensity will be removed}

\item{max_peak_num}{The maximum number of peaks to keep, set to -1 to disable}

\item{scale_int}{Normalized target value, set to -1 to disable}

\item{normalize_intensity}{Whether to normalize the intensity to sum to 1}

\item{precursorMz}{precursorMz of ms2. If spMat is a ms2, then it will usually have precursor ion, m/z greater than the precursorMz is unreasonable}
}
\value{
spMat.
}
\description{
Clean a spMat includes removing noise, merging similar peaks, and intensity normaliaztion.
}
\examples{
mz <- c(21.3300, 40.1320, 86.3400, 138.3290, 276.5710, 276.5830)
intensity <- c(100, 1300, 4030, 10000, 31600, 1000)
standardRow <- tibble::tibble(mz = list(mz), intensity = list(intensity))
spMat <- get_spMat(standardRow)
spMat <- clean_spMat(spMat)

}
