% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSpectra.R
\name{clean_spMat}
\alias{clean_spMat}
\title{clean_spMat}
\usage{
clean_spMat(
  spMat,
  tol_da2 = 0.02,
  tol_ppm2 = -1,
  min_mz = -1,
  max_mz = -1,
  noise_threshold = 0.05,
  max_peak_num = -1,
  scale_int = 100,
  normalize_intensity = FALSE
)
}
\arguments{
\item{spMat}{The matrix from the mass spectrum needs to contain two columns, mz and intensity.}

\item{tol_da2}{Under the Da unit, two peaks are considered the tolerance of one peak.}

\item{tol_ppm2}{Under the ppm unit, two peaks are considered the tolerance of one peak.}

\item{min_mz}{Minimum value of mz.}

\item{max_mz}{Maximum value of mz.}

\item{noise_threshold}{Peak intensity below noise_threshold * max(intensity) will be considered noise.}

\item{max_peak_num}{The number of peaks in the mass spectrum.}

\item{scale_int}{Normalized target value.}

\item{normalize_intensity}{Whether to normalize the intensity to sum to 1}
}
\value{
spMat.
}
\description{
Clean a spMat includes removing noise, merging similar peaks, and intensity normaliaztion.
}
\details{
The value of the parameter is set to -1 to turn off.
}
\examples{
mz <- c(21.3300, 40.1320, 86.3400, 138.3290, 276.5710, 276.5830)
intensity <- c(100, 1300, 4030, 10000, 31600, 1000)
standardRow <- tibble::tibble(mz = list(mz), intensity = list(intensity))
spMat <- get_spMat(standardRow)
spMat <- clean_spMat(spMat)
}
