% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSpectra.R
\name{compare_spMat_entropy}
\alias{compare_spMat_entropy}
\title{compare_spMat_entropy}
\usage{
compare_spMat_entropy(x, y, tol_da2 = 0.02, tol_ppm2 = -1)
}
\arguments{
\item{x}{Query spMat}

\item{y}{Library spMat}

\item{tol_da2}{Under the Da unit, two peaks are considered the tolerance of one peak.}

\item{tol_ppm2}{Under the ppm unit, two peaks are considered the tolerance of one peak.}
}
\value{
A score.
}
\description{
Compare two spMat using entropy
}
\examples{
mz <- c(21.3300, 40.1320, 86.3400, 138.3290, 276.5710, 276.5830)
intensity <- c(100, 1300, 4030, 10000, 31600, 1000)
standardRow1 <- tibble::tibble(mz = list(mz), intensity = list(intensity))
spMat1 <- get_spMat(standardRow1)
mz <- c(21.3000, 40.1120, 86.3200, 138.3210, 276.5310)
intensity <- c(100, 1300, 4030, 10000, 31600)
standardRow2 <- tibble::tibble(mz = list(mz), intensity = list(intensity))
spMat2 <- get_spMat(standardRow2)
Normalize intensity (sum to 1)
spMat1 <- clean_spMat(spMat1, normalize_intensity = TRUE)
spMat2 <- clean_spMat(spMat2, normalize_intensity = TRUE)
compare_spMat_entropy(x = spMat1, y = spMat2, tol_da2 = 0.05)
}
