% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_matrix.R
\name{compare_spMat_ndotproduct}
\alias{compare_spMat_ndotproduct}
\title{compare_spMat_ndotproduct}
\usage{
compare_spMat_ndotproduct(
  x,
  y,
  joinpeak = "inner",
  m = 0L,
  n = 0.5,
  ms2_tolerance_in_da = -1,
  ms2_tolerance_in_ppm = 5
)
}
\arguments{
\item{x}{spMat_query.}

\item{y}{spMat_lib.}

\item{joinpeak}{inner, left, right, outer; More details please see \link[Spectra:joinPeaks]{Spectra::joinPeaks}}

\item{m}{\code{numeric}, weighting for the first column of x and y ("m/z")}

\item{n}{\code{numeric}, weighting for the second column of x and y ("intensity")}

\item{ms2_tolerance_in_da}{The MS2 tolerance in Da, set to -1 to disable}

\item{ms2_tolerance_in_ppm}{The MS2 tolerance in ppm, set to -1 to disable}
}
\value{
A value.
}
\description{
The dot product method is used to compare the similarity of two mass spectrum matrices.
}
\examples{
mz <- c(21.3300, 40.1320, 86.3400, 138.3290, 276.5710, 276.5830)
intensity <- c(100, 1300, 4030, 10000, 31600, 1000)
standardRow1 <- tibble::tibble(mz = list(mz), intensity = list(intensity))
spMat1 <- get_spMat(standardRow1)
mz <- c(21.3000, 40.1120, 86.3200, 138.3210, 276.5310)
intensity <- c(100, 1300, 4030, 10000, 31600)
standardRow2 <- tibble::tibble(mz = list(mz), intensity = list(intensity))
spMat2 <- get_spMat(standardRow2)
# Spectra::joinPeaks(x = spMat1, y = spMat2, type = "inner", tolerance = 0, ppm = 200)
compare_spMat_ndotproduct(x = spMat1, y = spMat2, joinpeak = "outer", ms2_tolerance_in_da = 0.03)
}
